employeeApp/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.16
djangorestframework==3.12.1
idna==2.10
mysqlclient==2.0.1
numpy==1.18.5
panda==0.3.1
pandas==0.24.2
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
